+++ neoforge/src/main/java/net/mehvahdjukaar/sawmill/mixins/neoforge/RecipeManagerMixin.java
@@ -14,6 +14,8 @@ import net.minecraft.util.profiling.ProfilerFiller;
 import net.minecraft.world.item.crafting.RecipeHolder;
 import net.minecraft.world.item.crafting.RecipeManager;
 import net.minecraft.world.item.crafting.RecipeType;
+import net.minecraft.world.level.block.FireBlock;
+import net.neoforged.neoforge.common.crafting.IngredientType;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;