+++ changelog.md
@@ -1,1 +1,3 @@
-fix on neo that for some rason had some production only mixin issues. as usual
+- mod will now ignore all custom ingredients by default
+- this was done to prevent breaking said ingredients as they had to be initialized before tags were properly loaded
+- you might want to reset your configs for this to take effect